requirements.txt
absl-py==0.7.1
astor==0.7.1
atari-py==0.1.7
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
baselines==0.1.5
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
constantly==15.1.0
cycler==0.10.0
Cython==0.29.7
dill==0.2.9
docker-py==1.10.3
docker-pycreds==0.2.1
doom-py==0.0.15
fastzbarlight==0.0.14
future==0.17.1
gast==0.2.2
glfw==1.7.1
glob2==0.6
go-vncdriver==0.4.19
grpcio==1.20.0
gym==0.12.0
gym-pull==0.1.7
gym-retro==0.7.0
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
incremental==17.5.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
mpi4py==3.0.1
mujoco-py==2.0.2.2
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
pbr==5.1.3
Pillow==6.0.0
pkg-resources==0.0.0
ppaquette-gym-doom==0.0.6
progressbar2==3.39.3
protobuf==3.7.1
pycparser==2.19
pyglet==1.3.2
PyHamcrest==1.9.0
PyOpenGL==3.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.10.0
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
stable-baselines==2.5.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
Twisted==19.2.0
txaio==18.8.1
ujson==1.35
urllib3==1.24.2
websocket-client==0.56.0
Werkzeug==0.15.2
zmq==0.0.0
zope.interface==4.6.0
